feat(logout): add uselogout hook
